Further JSDoc
Started adding some more initial setup for JSDoc type checking! This allows you to get variable type checking on your code without moving to TypeScript. This can help you keep track of what kinds of values are going into your functions as parameters, and what kinds are coming out as return values, and the like. Using this can help you ensure that you are accessing properties and values on your variables correctly, because only a select few things are available in terms of things like dot notation, so only some kinds of variables have some features. Sorry if that was wordy, just thought it would be good to explain why it helps! Also, having a good editor like VSCode or something similar really helps in regards to this setup improving your development workflow, because you get these checks on your code right as you type it. If you try to add dot notation to a property of an object, but it's actually a Number at that point in the code, your editor will show you a 'typo'-looking error, similar to that of what something like Word or PowerPoint would do, for a word mispelling. It's very similar in that regard.