App Definition Types
Typed out the objects for each of the app definitions!

Also, found some helpful resources for this commit and the previous one, forgot to add the last one which added the support for the module types resolution, it helped out a lot here! Turns out `moduleResolution: Node` was required in the TSConfig for it to work also.

https://stackoverflow.com/questions/53792043/how-do-i-provide-tsc-with-typescript-definitions-for-uri-imports
https://stackoverflow.com/questions/64969300/typescript-index-signatures-in-mapped-type

There's a bunch of different things to type out, but I'm trying to do it in little stages, as I find things I know how to document. It's going well actually! It's going to make it much easier to migrate to full TypeScript eventually, as everything will already be documented out, it will just have to be a few VS Code refactoring steps to change the syntax over. I like doing it this way, compared to going all TypeScript all-in right away, as you area leaving the code in-place, and adding documentation to it, before you uproot it and add in a compilation step, and possibly break other things on accident if you don't expect the changes that might come with moving to modules, or things like that.

Having all of this typed out will also help a bunch with adding ES module support, so 'tree shaking' in the sense of the global scope will be very helpful for distinguishing between all of the variables between all of the apps' sources. That's one of my favorite things about modules, you can see where each thing is implemented respectively.

Music for the night:
https://www.youtube.com/watch?v=HPW5-U4dix8&list=PL63F40D2305D04496&index=7