Lib Types + Params
I think the biggest part that's making this difficult is the use of `document.getElementById()`, as you can't get any type inferences from that in regards as to what kind of element you are getting out of that is. And that it's not just a single reference to that element, but that it queries the same element over and over in multiple places, rather than just keeping a single variable stored for that element.

I'm wondering if it might actually be easier to hackily force-adjust the return types for `document.getElementById()` to instead return that of a known ID map for all the IDs on the site, as that will keep me from needing to add type assertions to every single one. I'm actually debating that now, it's a much quicker way of doing it, let's see if it helps.